refactored drugs section
