+++ client/src/app/views/diseases/diseases-detail/diseases-summary/diseases-summary.component.less
