+++ client/src/app/views/diseases/diseases-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { DiseasesDetailComponent } from './diseases-detail/diseases-detail.component';
 import { DiseasesHomePage } from './diseases-home/diseases-home.page';
-import { DiseasesSummaryComponent } from './diseases-summary/diseases-summary.component';
+import { DiseasesSummaryComponent } from './diseases-detail/diseases-summary/diseases-summary.component';
 
 import { DiseasesComponent } from './diseases.component';
