+++ client/src/app/views/drugs/drugs-detail/drugs-detail.component.html
@@ -1,25 +1,27 @@
-<ng-container *ngIf="drug$ | ngrxPush as drug" >
+<ng-container *ngIf="drug$ | ngrxPush as drug">
   <cvc-section-navigation [displayName]="drug.name"> </cvc-section-navigation>
   <nz-page-header class="site-page-header">
     <nz-page-header-title>
-      <nz-space nzDirection="horizontal">
-        <i nz-icon nzType="civic:intervention" *nzSpaceItem></i>
-        <span *nzSpaceItem>{{ drug.name }}</span>
-      </nz-space>
+      <i nz-icon
+        nzType="civic:intervention"></i>
+      {{ drug.name }}
     </nz-page-header-title>
     <nz-page-header-extra *ngIf="drug.ncitId">
-      <cvc-link-tag [href]="drug.drugUrl" tooltip="View in the NCIthesaurus">
+      <cvc-link-tag [href]="drug.drugUrl"
+        tooltip="View in the NCIthesaurus">
         {{ drug.ncitId }}
       </cvc-link-tag>
     </nz-page-header-extra>
     <nz-page-header-content>
       <div class="content">
-      <nz-descriptions nzSize="small"
-        [nzColumn]="1"
-        nzBordered="true">
-        <nz-descriptions-item *ngIf="drug.drugAliases.length > 0" nzTitle="Aliases" nzSpan="2">
-          {{ drug.drugAliases.join(', ') }}
-        </nz-descriptions-item>
+        <nz-descriptions nzSize="small"
+          [nzColumn]="1"
+          nzBordered="true">
+          <nz-descriptions-item *ngIf="drug.drugAliases.length > 0"
+            nzTitle="Aliases"
+            nzSpan="2">
+            {{ drug.drugAliases.join(', ') }}
+          </nz-descriptions-item>
         </nz-descriptions>
         <router-outlet></router-outlet>
       </div>
