+++ client/src/app/views/drugs/drugs-detail/drugs-summary/drugs-summary.component.less
