+++ client/src/app/views/drugs/drugs-home/drugs-home.page.html
@@ -1,12 +1,14 @@
+<cvc-section-navigation> </cvc-section-navigation>
 <nz-page-header class="site-page-header">
   <nz-page-header-title>
-    <i nz-icon nzType="civic:intervention"></i>
-    Drugs
+    <i nz-icon
+      nzType="civic:intervention"></i> Drugs
   </nz-page-header-title>
 
   <nz-page-header-content>
     <div nz-row>
-      <div nz-col [nzSpan]="24">
+      <div nz-col
+        [nzSpan]="24">
         <div class="content">
           <cvc-drugs-table></cvc-drugs-table>
         </div>
